[alloc @ alla@peach-20060703151324-0c846d6408654d63] Removed personIDonly filter from taskSummary page, fixed task filter to work properly for Assigned To searches.

+++ util/misc/update_alloc_dev_database.sh
@@ -5,16 +5,14 @@ function quit {
   exit;
 }
 
-[ -f /tmp/person.sql ] && quit "Pls remove /tmp/person.sql"
+[ -z "${1}" ] && die "usage: ${0} FILE.sql"
 
-FILE="alloc_dev.sql"
+FILE="alloc_test.sql"
 rm -f ${FILE}
 
-echo "drop database if exists alloc_dev; create database alloc_dev; use alloc_dev;" > ${FILE}
+echo "drop database if exists alloc_test; create database alloc_test; use alloc_test;" > ${FILE}
 
-echo "Enter user 'alloc' mysql pass: "
-mysqldump -u alloc -p alloc | grep -v "INSERT INTO active_sessions" >> ${FILE}
-[ "${?}" -ne "0" ] && quit "Could not dump out alloc database!"
+cat ${1} >> ${FILE}
 
 
 # append updates to the SQL file.
@@ -38,14 +36,10 @@ delete from eventFilter;
 " >> ${FILE}
 
 
-#cat ../db_changelog.sql >> ${FILE}
-#[ "${?}" -ne "0" ] && quit "Trouble concatenating db_changelog.sql file"
-
-
 
 # refresh alloc_dev database with this script.
 echo "Enter root mysql pass: "
-mysql -u root -p < ${FILE}
+mysql -u root < ${FILE}
 [ "${?}" -ne "0" ] && quit "Could not execute all of ${FILE}!"
 
 [ -f "${FILE}" ] && rm ${FILE}
